Improved persistence of path to comics
Previously if the user entered a path argument, it would only be
recognised at that specific instance of the project. Starting the project
another time without providing the path argument would result in the
program not acknowledgin the previous comics.

A file named 'pathfile' (stored at 'C:\\XKCD' or the value set at
the variable: 'defaultpath') is used to store the path to the comics.

This enables the project to provide some form of persistence in knowing
where the comics where stored at the previous instances it was run.